v 1.0 - Rust + Linux fingerprint spoofing engine
I don't think I've ever written one of these at a reasonable hour :):

mitmproxy was really fighting me, so I decided to just reimplement a lot of their functionality into a hand-built rust proxy. There's like, a shit ton of documentation, have fun!

It does almost everything. I fixed all the JS problems. It could be better, but alas.

This is it. This is what it all culminates into. Packaging and UI next!